# backend/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Farmasi Queue API",
    "version": "1.0.0",
    "description": "API documentation for Antrian Farmasi system"
  },
  "host": "192.168.6.106:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/queue/GMCB": {
      "post": {
        "tags": [
          "Queue"
        ],
        "summary": "Create Antrian GMCB",
        "description": "Create Antrian GMCB",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "medicineType": {
                  "example": "any"
                },
                "lokasi": {
                  "example": "any"
                },
                "origin": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/create-farmasi/GMCB": {
      "post": {
        "tags": [
          "Queue"
        ],
        "summary": "Create Antrian GMCB",
        "description": "Create Antrian GMCB",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "medicineType": {
                  "example": "any"
                },
                "lokasi": {
                  "example": "any"
                },
                "origin": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/GMCB-temp/": {
      "post": {
        "tags": [
          "GMCB TEMP"
        ],
        "summary": "Create a new appointment",
        "description": "Creates a new appointment",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "lokasi": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "GMCB TEMP"
        ],
        "summary": "Get All Appointment",
        "description": "Get All Appointment",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-temp/get-list": {
      "post": {
        "tags": [
          "GMCB TEMP"
        ],
        "summary": "Get Farmasi List",
        "description": "Get Farmasi List",
        "parameters": [
          {
            "name": "x-cons-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-timestamp",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-signature",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "payload": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-temp/antrian": {
      "get": {
        "tags": [
          "GMCB TEMP"
        ],
        "summary": "Get Antrian",
        "description": "Get Antrian",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-temp/{id}": {
      "get": {
        "tags": [
          "GMCB TEMP"
        ],
        "summary": "Get By NOP",
        "description": "Get By NOP",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "GMCB TEMP"
        ],
        "summary": "Delete By NOP",
        "description": "Delete By NOP",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-temp/update": {
      "patch": {
        "tags": [
          "GMCB TEMP"
        ],
        "summary": "Update By NOP",
        "description": "Update By NOP",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "queue_status": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-temp/{category}": {
      "get": {
        "tags": [
          "GMCB TEMP"
        ],
        "summary": "Get By Category",
        "description": "Get By Category",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-temp/verify": {
      "post": {
        "tags": [
          "GMCB TEMP"
        ],
        "summary": "Verify Temp Queue",
        "description": "Verify Temp Queue",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "tempId": {
                  "example": "any"
                },
                "verificationData": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}